I'm really tired of thinking of the commit message, so i just push things I made...

+++ src/main/java/com/onxshield/invoiceyou/invoicestatement/service/invoiceService.java
@@ -84,7 +84,7 @@ public class invoiceService {
         int generatedCount = 0;
         List<String> categories = Arrays.asList(requestDetails.getCategories().split(", "));
         List<inventory> allProducts = inventoryRepository.findAll();
-        long targetedTotal = 0L;
+        double targetedTotal = 0d;
         while(targetedTotal<=requestDetails.getTotalTTC()+1.5 & targetedTotal>=requestDetails.getTotalTTC()+1){
 
         }
